HSEARCH-3442 Fix configuration property retrieval to properly trim strings
We expect property source to get their data from property files in
particular, where extra spaces would be common.
And, at least for the current property types, leading or trailing spaces
do not have any meaning, so we might as well remove them automatically.

Note that we used to do this in Search 5 too; see
ConfigurationParseHelper.