HSEARCH-3446 Rename value converters ((To|From)Index.*ValueConverter) to use the word "document" instead of "index"
Using "index" is particularly confusing for the ID value converter,
because it makes it seem like the identifier identifies an index.
And if we change that one, we may as well be consistent and change
the others.

This change should not be a big deal, because we only used "Index" in
the name to avoid confusion with *entity* fields. "Document" should work
just as well.