HSEARCH-3446 Remove UserDocumentFieldConverter, ElasticsearchFieldConverter and LuceneFieldConverter
The UserIndexFieldConverter class seem unnecessary: it aggregates two
converters, which is not really useful now that converters are always
used from within a factory, knowing that each factory only uses one of
the two aggregated converters.

The LuceneFieldConverter/ElasticsearchFieldConverter seem unnecessary:
some converters add behavior to convert from the exposed type to the
encoded type (e.g. LocalDate to Long), but that's something that we
already do in the codec class when indexing. Why duplicate the
conversion code?

Note that we have to duplicate some code in order to make everything
work, but this will be solved in the next commits.