HSEARCH-3453 Make sure to execute projection operations that potentially rely on the session in the user thread
Most of this commit is only side-effects; the purpose of this commit is
the changes in ElasticsearchFieldProjection and LuceneFieldProjection.

It is possible that a user projection converter makes use of
HibernateOrmExtension to gain access to the session, in particular.
Since session-related operations may be blocking, let's make sure
to execute them in the user thread if possible.