HSEARCH-3291 Add to LuceneGeoPointFieldCodec the ability to work with any field path (not just the one passed to its constructor)
Previously LuceneGeoPointFieldCodec made some assumption that the
absoluteFieldPath parameter would always be the same as the one passed
to its constructor.

This assumption currently holds, but only by happenstance. In the future
we may share a codec between multiple fields, either because of the
introduction of dynamic fields or simply because the field types we
are about to introduce are going to be re-used by bridge implementors.

More importantly, the field predicate/sort/projection builder factories
currently hold a reference to the codec, and *they* are supposed to be
usable for any field path (they take the field path as a parameter). So
it does not make much sense for the codec to be only usable for one
field path.

The assumption was there to optimize performance, but right now it does
more wrong than right, so I decided to remove it for now. We may
re-introduce it later.

I tried to avoid unnecessary Set instantiation where possible, but
String concatenation cannot be avoided at the moment. We may be able to
avoid it using the following pattern in the future:

public interface LuceneIndexFieldCodec<F, M> {
  M createMetadata(String absoluteField);
  void encore(LuceneDocumentBuilder documentBuilder, M fieldMetadata, F value);
}

... and storing the field metadata somewhere to re-use it.

Alternatively we could decide to make the field predicate/sort/projection
builder factories AND the codecs field-specific. We will have to see
what's best.