HSEARCH-3052 Only deal with one type mapping per index in Elasticsearch index administration code
We do not need more, since we are only ever interested in one type per
index. Also, starting from ES6 there simply cannot be more than one type
per index.

Note that "_doc" will be the only allowed type name in Elasticsearch 7,
and the concept of type name will be removed in Elasticsearch 8 (but the
endpoints will still include a "_doc" path component).
We can't start type names with an underscore in Elasticsearch 5.6,
though, so we cannot switch to "_doc" as the default type name yet.