HSEARCH-3524 Use the Hibernate ORM Metadata instead of the Metamodel to extract metadata
The Hibernate ORM Metamodel is built very late in the ORM bootstrapping
process, so late actually that ORM may have sent some requests over the
network before it is built (for second level caches in particular).

As a result, Quarkus cannot build a Metamodel and provide it to us
during the "compile-time boot". But we want Hibernate Search to start
booting during this "compile-time boot", because we need reflection and
it's much easier to use reflection then than at runtime, due to how
Quarkus/SubstrateVM work.

Thus we need to extract whatever information we can from Metadata, which
is a more "raw" version of the Metamodel.