HSEARCH-3080 Use "doc" as type mapping name in ES5 and ES6
Facts:

1. We cannot start a type name with an underscore in ES5.6.
2. In ES7 and above, we will be able to use "_doc" instead of the type
name, and it will work irrespective of the actual type name.
3. In ES7, using a type name in APIs will be deprecated, and in ES9 it
will not be possible anymore.

We want user migration from 5.6 to 6, then 6 to 7 to be as painless as
possible:

1. If they have external code manipulating the indexes (triggering
searches through the REST APIs, for example), this code should require
as little change as possible when they upgrade.
2. If possible, they should not have to reindex when they upgrade.

Using "_doc" as the type name everywhere would give us a consistent
behavior for all versions of Elasticsearch, but it is not possible,
because 5.6 does not allow such name.

If we use "doc" in ES5/6, then "_doc" in ES7+:

* In ES5, users will be able to send requests natively (using a REST
client) with the "doc" type name.
* When migrating from ES5 to ES6, the type name will not change.
* In ES6, users will still be able to send requests natively with
the "doc" type name.
* When migrating from ES6 to ES7, the type name will not change.
* In ES7, users will still be able to send requests natively with
the "doc" type name. Additionally, they will be able to send
requests using "_doc" instead of "doc".
* When migrating from ES7 to ES8, the type name will not change.
* In ES8, users will still be able to send requests natively with
the "doc" type name *if* they add the request parameter
"?include_type_name=false". Additionally, they will be able to send
requests using "_doc" instead of "doc".
* When migrating from ES8 to ES9, the type name will be removed
completely.
* In ES9, users will have to send native requests using "_doc"
instead of "doc".