HSEARCH-3089 Make the default analyzer in ScopedAnalyzer a constant
For now we do not offer a way to customize it, so we may as well avoid
instantiating it multiple times.

If we ever define a default analyzer, it will have to be set in the
field metadata, so that it can be used when building a match predicate,
for example. So we will have an entry in the scoped analyzer map,
and the default in the scoped analyzer will not matter.