HSEARCH-3511 Move SearchQuery (the engine type) to SPI
... so that it's clear that mappers need to declare their own type,
wrapping the engine-provided query. This is the safest design: it allows
to evolve SPIs independently from mapper APIs, avoiding naming conflicts
in particular.

If we want to define a common super-interface for the query type of all
mappers, we can always do that later, but in the meantime let's be safe.