HSEARCH-3403 Remove the inheritance from our Session/Query APIs to Hibernate ORM APIs
Instead, allow to convert objects from our APIs to JPA/ORM APIs through
explicit calls to toXXX() methods.

There are two advantages to this:

1. Auto completion will be clearer: by default, only Search methods will
be displayed, and we won't clutter the suggestions with the dozens of
methods of the JPA/ORM APIs.
2. We will be able to explain clearly on the javadoc of the conversion
methods that the converted objects only implement a subset of the
JPA/ORM APIs. And we will have more legitimacy in explaining that, since
those objects are not our primary APIs.