HSEARCH-3070 Add a concept of "role" to register beans to use in a specific extension point
This will be useful for the Elasticsearch client in particular, to allow
registration of HTTP client configurers by external modules, and then
allow the Elasticsearch backend to retrieve all these configurers and
use them.

Relying on Java services would not be flexible enough: users would not be
able to register Spring or CDI beans, or even just beans that were
instantiated directly in their application.

Relying on Spring or CDI directly is not an option either, because our
SPIs are a bit limited and do not allow to retrieve *multiple* beans
matching a given type (they will fail if multiple beans match a single
reference).

Roles force us to register beans explicitly, which is a bit annoying,
but at least they solve the problems we had with the two solutions
above.