HSEARCH-3070 Remove the concept of ServiceManager
Currently we do not need services, and I don't see any reason we should
need them in the near future.

Should we ever need services, the BeanProvider will be the most obvious
place to implement this feature.

The "role" feature can already be used to get a set of beans with an
assigned role.
For services, we could use roles and expect only one bean to be
assigned a given role. That's more or less equivalent to what we already
have in Search 5, where having multiple implementations of a same
service will simply fail.

If we ever need to retrieve the exact same instance of the service
multiple times from the bean provider, or to provide a way to resolve
conflicts when there are multiple candidates for a single service,
I'm confident we could implement this fairly easily.
But as I said, currently it's just not necessary.