HSEARCH-2464 Optimize implementation of "exists" predicate for Lucene
When possible, avoid indexing field names and rely on doc values or
field norms instead.

This matches the defaults of Elasticsearch.

There is one downside which I think is acceptable: "exists" queries no
longer work on unknown fields or on fields that have a different type in
the targeted indexes. Working around this while keeping the optimization
would require to perform queries differently: essentially we would need
to build one query per index, run the queries separately on each index,
and then combine the results. I'm not sure how complex this would be, so
I won't try for now. Let's keep it as an improvement for later.

We could add an option to disable indexing of field names completely,
but I do not think that would bring much: currently field names are only
indexed for numeric fields that are not sortable and string fields that
do not have norms.