HSEARCH-3549 Format MonthDay with a year forced to 0 in Elasticsearch
We can't format MonthDay as '--MM-dd' because that would set the year
to 1970 in Elasticsearch, and as a result february 29th would be
rejected.

Using the format 'uuuu-MM-dd' and setting a leap year ourselves allows
us to work around the problem, even though users attempting to index
data will have to take care to use the exact same year.