HSEARCH-3541 Do not rely on the Elasticsearch client's "max retry timeout" anymore
It is no longer available in Elasticsearch 7, and we already have our
own timeout. It just won't retry anymore.
See https://github.com/elastic/elasticsearch/issues/25951#issuecomment-460929775

Note this clears the path for more useful timeouts that start when the
request is sent to the cluster rather than when the request is queued.
See https://hibernate.atlassian.net/browse/HSEARCH-2836.