HSEARCH-3316 Determine whether Elasticsearch works should trigger a refresh depending on the work plan instead of the orchestrator
This will allow us to configure refreshes from the mapper.

The setting "force refresh after write" used to be determined by the
work orchestrator.

This caused two problems:

* It was not possible for the *mapper* to decide whether refreshes
should be forced or not, which is annoying because forcing refreshes
has an impact on throughput, and so do other settings in the mapper.
Ideally we'd want to configure everything at the mapper level, for the
sake of simplicity.
* It was not possible to change the setting from one changeset to
another, which is annoying because some users may want to leave it
disabled in general (for performance) and enable it on an opt-in basis.

With this commit, the setting is still sourced from index configuration
properties, but is determined by the work plan.
Since the work plan is created upon mapper request, this will allow us,
in the next commits, to source the setting from the mapper instead of
index configuration properties, which will allow us to solve the two
problems mentioned above.