HSEARCH-3544 Remove any internal state from SearchQueryContextImpl
... and rely exclusively on the state held by the query builder.

This will allow us to create separate contexts, for example implementing
a Lucene-specific interface, that point to the same query builder and
can create the query directly instead of the original context.

Note the removed comments in toQuery() were obsolete: they dated back to
a time where we had a "fully fluid" API, i.e. we allowed to do something
like:

.search(Book.class).predicate().match().onField(...).matching(...).toQuery()

... which lead to lots of corner casses like the one mentioned in this
comment, which is why we gave up on this syntax.