HSEARCH-3544 Allow to use backend-specific predicates and sort directly when extending the query context
This required to add generic type parameters to contexts, which is
rather unfortunate, but I couldn't find a better way to do it.

Initially the plan was to override methods in the backend-specific
interfaces, e.g. this method from `SearchQueryContext`:

```
S sort(Consumer<? super SearchContainerContext> sortContributor);
```

would be overridden this way in `LuceneSearchQueryContext`:

```
S sort(Consumer<? super LuceneSearchContainerContext> sortContributor);
```

In theory that makes perfect sense, because
`Consumer<? super LuceneSearchContainerContext>` is a supertype of
`Consumer<? super SearchContainerContext>`,
and a small test allows to check that:

```
Consumer<? super SearchContainerContext> var1 = null;
Consumer<? super LuceneSearchContainerContext> var2 = var1;
```

The above compiles just fine.

However, it seems the java compiler doesn't allow this kind of method
override, and we get a compilation error related to conflicting method
signatures.

So we can't do this, and have to rely on generic type parameters in
contexts.