HSEARCH-3353 Remove LuceneSearcher and split its logic between orchestrator and work
The "explain" work will have to share the orchestrator logic, but not
the work logic, so it will be cleaner that way.
Also, this makes the query works more consistent with the index works:
the execution context contains the index writer for index works, and the
index reader for query works.