HSEARCH-3463 Remove BuiltinContainerExtractor.AUTOMATIC
It lost its meaning with the introduction of @ContainerExtraction, which
is able to interpret a empty extractor array differently depending on
the value of @ContainerExtraction.extract. In particular it's able to
interepret that empty array as "use default extractors" when extract =
ContainerExtract.DEFAULT, which is what
BuiltinContainerExtractor.AUTOMATIC was created for in the first place.