HSEARCH-3314 Reorganize the various *BackendContext classes
In short, we define callback interfaces in packages that require
access to backend context, and we implement these interfaces in a single
class that gets passed down from the backend to wherever it's necessary.

This allows clean inter-package dependencies, like the previous
solution, but the new solution will be more suitable when we start
adding index writing features to the IndexScope interface:
we'll need to access multiple such interfaces in the index scope,
not just the one about search.