HSEARCH-3203 Move around the wildcard in EntityLoader generics for easier implementations
This does not change much overall, but it allows EntityLoader
implementations to use the actual loaded type as their generic
parameter, instead of a supertype like they used to.
This will be useful when introducing more complex EntityLoader
implementations that need to manipulate other objects that refer to that
generic parameter.