HSEARCH-3349 Always execute a single statement when loading an entity whose document ID is not the entity ID
The old code used to execute a first SQL statement to retrieve entity
proxies, then additional statements to initialize the proxies. When
no batch size was set, we could end up executing N + 1 queries.

The new code apparently loads the entities directly as part of the first
statement, so it should be better. We'll add tests in the next commits.