HSEARCH-3349 Do not expect any order for index names in the stub backend
In some situations, such as when a query targets an abstract type, the
order of index names passed to the backend is deterministic, but not
easily predictable, because it depends on the execution order of some
mapping operations at boot time. So it's easier to just ignore the
order, which isn't very important anyway.