HSEARCH-3609 Move projection conversion for ValueBridges to the bridge itself
... instead of requiring the bridge to implement bind().

We're about to move the bind() feature to another, more advanced class
(the bridge builder). This means setting the projection converter
explicitly will not be possible anymore without using a bridge builder.

But projection convertion is a very basic feature, which should
definitely be implementable simply by defining a ValueBridge, without
having to mess with bridge builders.

By adding a fromIndexedValue() method to ValueBridge, we solve this
exact problem.