HSEARCH-3609 Remove support for direct bridge reference for type, property and routing key bridges
The type bridge and property bridge are worthless without a declared
index field, and index fields can only be declared by the binder. So it
doesn't make sense to reference a bridge directly without a binder.

Routing key bridges could theoretically be used without a binder, but
only if they rely exclusively on the ID and entity type to generate the
routing key. It's doubtful such a simple bridge would be useful,
and routing key bridges are advanced usage anyway, so we'll remove
support for direct bridge reference for these too, so we're consistent
with type and property bridges.