HSEARCH-3314 Remove references to index name and routing key in Lucene works
This information is not used in the work itself, but in the component
that creates the work, to route the work to the appropriate executor.
So the work has no control over where it will be routed, and thus having
this information in the work itself is useless at best, misleading at
worst.