HSEARCH-3642 Introduce a single enum for configuration of field conversion
We need this because, when defining a range aggregation, we need to
convert both ways: from the given ranges to ranges of the type of the
index field, and from the returned ranges to ranges of the given type.

The enum is structured and named so as to be consistent with
ContainerExtract.