HSEARCH-3441 Remove the "on"/"or" prefixes from "field"/"fields" methods in the predicate DSL
For consistency with the projection, sort and aggregation DSLs.

One big advantage is that calls to .field() no longer give the
impression that the predicate will match if *any* field matches.
This is currently true, but there are plans to provide options to tune
this: https://hibernate.atlassian.net/browse/HSEARCH-3327