HSEARCH-3684 Fix deeply nested indexed-embedded fields being incorrectly included
I ended up changing the approach completely: instead of pre-computing
the set of included paths when we compose filters, the filters simply
keep a reference to their parent and compose their decision "dynamically"
each time "isPathIncluded()" is called.

This is probably less efficient, but it's far easier to implement, debug
and understand. And as it turns out, it also seems to fix the bug.