HSEARCH-3684 Move path tracking for indexed-embedded to a dedicated class
... so that we'll be able (in the next commits) to share path tracking
state across all "uses" of the same indexed-embedded.

This is especially important when an indexed-embedded is used
recursively: in that case, the deepest level may not make use of all
includePaths, and if we don't share tracking state, it will incorrectly
deem some includePaths useless.

Run IndexedEmbeddedIncludePathsAndDepthIT before this commit for an
example of such false positive.