HSEARCH-3084 Pre-stop backends and index managers in parallel
Mainly, this means that on shutdown, we'll stop accepting works
for all backends/index managers immediately, and *then* we'll wait for
ongoing works to complete.

The improvement over the previous behavior will probably be negligible,
but what's more important is that the path towards an API for explicitly
starting/pre-stopping/stopping index managers is now clearer.