HSEARCH-1375 Correctly report uncommitted works to the failure handler upon Lucene indexing failures
We used to only report works from the current workset after the one that
failed, but really the uncommitted works before the failing work may
also end up not being applied to the index, especially in case of commit
failure.

For example if we execute:

1. Workset 1: work 1, work 2 (commit strategy: no commit)
2. Workset 2: work 3, work 4, work 5 (commit strategy: no commit)

... and work 4 fails, then works 4 and 5 will obviously not be applied,
but work 1, 2, and 3 may end up not being applied either, depending on
the error, because we didn't commit them before the failure.