HSEARCH-1375 Correctly report failures and uncommitted works to the failure handler upon Elasticsearch indexing failures
1. Report one failure per failed work. We used to report one failure
with the first failed work as the "failing operation" and all other
works as simply "uncommitted operations", but that's wrong: in some
cases (bulk work failure) there are multiple failing works and the first
one is not more responsible of the failure than the others.
2. Blame uncommitted works on the first failed work in each sequence.
This is necessary now that we may report multiple failures per sequence.