HSEARCH-3748 Remove the bridge-based programmatic mapping APIs for TypeBridge, PropertyBridge and RoutingKeyBridge
I.e. the APIs where we don't specify a binder, but a bridge directly.

These APIs are not documented, not tested, but more importantly they
are pointless: defining a brige this way simply cannot work, because
there would be no binder to define the dependencies of the bridge,
so Hibernate Search would simply throw an exception on bootstrap
because of the unknown dependencies.

Even more critical, for the TypeBridge and PropertyBridge, applying a
bridge without a binder would mean that the bridge does not have any
index field reference to work with, so it effectively cannot write to
the index.