HSEARCH-3797 Move various low-level Lucene-related classes to the "lowlevel" package
So that they can be used from other packages without introducing
circular dependencies between packages.

We will need this in the following commits, though we will only
use a few of these classes.

The new location makes sense regardless of this ticket, because these
classes really do provide low-level operations, and NestedDocsProvider
is not strictly related to field types.