HSEARCH-3786 Simplify the collection of child documents in the Lucene backend
We still have the same problem as before (explained in HSEARCH-3786):
children are collected for all matching documents, not just topdocs,
which is ridiculously inefficient.

However, we will be fixing this problem in the next commits for
HSEARCH-3797 by differentiating collectors that must be applied to
all matching documents and collectors that must be applied to topdocs
only.