HSEARCH-3792 Specify a format when requesting doc values fields in ES 6.7/6.8
So that we avoid the following warning:

There are doc-value fields which are not using a format.
The output will change in 7.0 when doc value fields get formatted based on mappings by default.
It is recommended to pass [format=use_field_mapping] with a doc value field
in order to opt in for the future behaviour and ease the migration to 7.0

Theoretically we should do this for 6.6 as well, and perhaps also 6.5
and 6.4, since that parameter was introduced in 6.4 and is known to
trigger warnings at least in 6.6 (maybe 6.5 and 6.4 too).
But I think making sure the latest 6.x works correctly is enough.