HSEARCH-3775 Implement forced refreshes in LuceneWriteWorkProcessor
Essentially, after this change, it is possible to write to the index and
force the change to be visible to search queries without requesting a
commit.

Before this change, we would systematically commit in order to make the
changes visible, which 1) is overkill and 2) is not effective when a
refresh_interval different from 0 is configured.