HSEARCH-3824 Add automatic filtering by routing keys to search queries
Previously, we used to direct search queries to the shards corresponding
to the given routing keys, but we didn't filter search results
accordingly, so the results could include documents that were indexed
with a different routing key that happens to be assigned to the same
shard. The user was expected to add filtering on their own.

With this change, users no longer have to add filtering on their own.