HSEARCH-3103 Rename MultiValue to SortMode
Since we always pass it to methods expecting a "mode", it would be weird
for this class to not even mention the term "mode".
Also, it's only used in sorts, and if we even introduce a similar
setting for aggregations, we should use a different enum because the
values we accept may be different (in particular, we may accept "NONE").