HSEARCH-3822 Do not track uncommitted works in LuceneWriteWorkProcessor
It's costly and very complex to implement if we start accepting concurrent
commits of the index writer, like we will soon.

One consequence is that we no longer report the exact list of uncommitted
works upon low-level write failure.
This should not be a problem in general as that kind of failure is
unlikely to be recovered from anyway.
Note that we didn't use to report uncommitted works upon low-level write
failure in Search 5: this was added in Search 6.