HSEARCH-3822 Move the responsibility of forced commit/refresh from LuceneWriteWorkProcessor to LuceneIndexIndexingPlan
This is necessary if we want to have multiple work queues and threads
per index, sharded by document ID. In that case, works from a single
work plan may be executed in different threads, thus we can no longer
expect commits and refresh to be executed "just after the works".
We need something to be notified that all works have been executed, and
to act accordingly.