HSEARCH-3872 Remove the ability to execute bulkable works outside of a bulk
Just execute them in a bulk, always. It's simpler, and more importantly
it allows us to get rid of the end-of-sequence refresh.

That end-of-sequence refresh must disappear if we want to have multiple
work queues and threads per index, sharded by document ID.
In that case, works from a single work plan may be executed in different
threads, thus we can no longer expect the refresh to be executed
"just after the sequence".

For now we'll execute refreshes as part of the bulk works, which may
incur a significant performance cost, as we may execute multiple
refreshes per index for each work plan.
We may want to look into this in the next commits, but since forced
refreshes already perform very badly, this is not a priority.