HSEARCH-3103 Disable sort tests in a very specific case for ES 5.6
For some reason, ES 5.6 seems to evaluate the max to 0 when a nested document
has a field with only negative floating-point values.
This causes problems in our tests relative to field sorts
because it brings a max from -42 to 0, which changes the relative order of documents.
This is most likely a bug, though I couldn't find the bug report or fix,
and it is fixed in ES 6.x.
Since 5.6 is really old and EOL'd anyway, it's unlikely to ever get a fix.
We'll just ignore tests that fail because of this.