HSEARCH-3822 Remove the concept of "work set" in the engine and Lucene backend
Work sets no longer make sense, at least during work orchestration,
since we submit works one by one without any information about the
relationship between works.

The "BatchedWork" is what replaces the concept of "work set". A "batched
work" is just a single work, not multiple works like the "work set", but
similarly to the work set it is responsible for submitting the underlying
work to the work processor and forwarding the execution result to the
submitter.