HSEARCH-3822 Add a lock to Lucene commits
... to avoid problems when closing the directory (e.g. to drop the
index) while a delayed commit is executing. Lucene throws an exception
when one tries to close an index writer as a commit is in progress,
so we need our own synchronization to make sure we're not committing
when we close the index writer.

Note the performance impact should be minimal as there is already a lock
on commit in the Lucene index writer.