HSEARCH-3872 Reduce the default max bulk size to 100
There seems to be a limit of around 250MB for all requests executing at
a given time, and Elasticsearch just rejects requests past that point.

With a parallelism of 10 and a max bulk size of 250, we can reach a
total of 2500 documents being indexed in parallel, which is apparently
enough to go over the 250MB limit in some tests with big documents
(around 100KB each, I suppose).

For the record, the test that highlighted the problem was
org.hibernate.search.integrationtest.backend.tck.search.query.SearchQueryTimeoutIT.