HSEARCH-3892 Avoid the use of ConjunctionDISI in TextMultiValueFacetCounts
First because ConjunctionDISI is internal to Lucene and shouldn't be
used elsewhere, and second because it won't work when we switch to our
own TextValues which do not extend DocIdSetIterator (nextDoc(),
advance(), etc.).

The performance cost shouldn't be too high as long as documents without
a value are relatively rare, and in any case we already pay the same cost
for numeric aggregations.