HSEARCH-3958 Break down MatchSearchPredicateIT into multiple reusable tests
So that we can avoid re-implementing tests for every predicate, in
particular for common features such as boosts.

Incidentally, this also results in a better organization for tests, as
we no longer have a 1700-line-long test class, but several 300-line-long
abstract classes regrouping only closely related tests.

Also, this changes how tests are initialized, so that we do much less
schema operations, which should speed up tests on Elasticsearch.