HSEARCH-3323 Fetch the top docs for the first 4 pages in the first .next() call of Lucene scrolls
We used to fetch the first 100 elements in Search 5, but I think this is
closer to what we want.

When using scrolling, we're unlikely to only want one page. Fetching
only the first page, then the first and second, then finally 1-4 when
we reach the 3rd page, seems like a waste of resources since we'll
likely always reach the 3rd page.

We might want to refine this behavior later, but for now it's close
enough.