HSEARCH-3108 Allow routing bridges to provide previous routes so that we can delete documents there
This effectively allows routing bridges to work on mutable properties,
e.g. a "status" property that would switch from "DRAFT" (not indexed)
to "PUBLISHED" (indexed) to "ARCHIVED" (not indexed).

Performance-wise, we will often send unnecessary delete commands to
the backend, but that's necessary to ensure the index is in sync.
We will avoid processing entities that are no longer indexed (like
the archived one in the example above).
Also, the unnecessary delete commands will target non-existing
documents; maybe we can hope that those commands are quicker to
execute than actual deletions?