HSEARCH-3282 Remove support for creating queries from parameters of incorrect type on numeric fields in the migration helper
We used to allow users to pass any value that could be converted to a
numeric value (e.g. Date, Calendar), and we converted that to the
corresponding numeric type (e.g. long). If that type matched the
internal type of the field, everything worked as intended.

This is no longer possible with Search 6, as the internal type of fields
changed in some cases (date/time types in particular) and is now
higher-level (e.g. Date => Instant, not long).