HSEARCH-4019 Rig scoring in SearchQueryFetchIT so that even sharded ES indexes yield the same sort
Sharded indexes would usually yield less-than-perfect scoring when it
comes to repeated terms, because the impact of term repetition on
scoring depends on the frequency of that term in a shard, and so it
depends on to which shard each document was routed, which is not easily
predictable.

By assigning a different term and boost to each document, we ensure that
the documents will be in the correct order.