HSEARCH-3999 Share more code between PojoIndexedTypeIndexingPlan and PojoContainedTypeIndexingPlan
Automatic indexing is quite hard to test, because there are lots of
possible combinations. We already have lots of tests, but I think we
missed some code paths because we mostly test indexed types, and more
rarely contained types.

As I don't think it would be reasonable to duplicate existing tests to
also cover contained types, this is the alternative: by sharing more
code between contained types and indexed types, we ensure that if it's
tested for one, it's likely to work for the other, too.