HSEARCH-4047 Return object fields as well when calling *SearchIndexesContext#field
This allows us to give more sensible error messages to the user when
they attempt to use a predicate/sort/projection that is not available on
object fields: instead of telling them the field does not exist, we tell
them the predicate/sort/projection is not available on object fields.

This also introduces *SearchObjectFieldContext, similar to *SearchValueFieldContext
but for object fields.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>