HSEARCH-4096 Resolve beans from built-in bean definitions first, then from the bean provider
That way, if the bean provider is buggy, a user who doesn't rely on
custom beans won't be affected.

That being said, this means we will introduce backward-incompatible
changes if we ever add new built-in bean definitions for an existing
extension point (e.g. a new layout strategy): users who picked the exact
same name for their own implementation will have it replaced with the
built-in one.

The next commits should offer a workaround.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>