HSEARCH-4115 Make sure to reset TestRules when retrying tests
We used to always fail the second and third attempt because
we hadn't reset Wiremock request expectations.

By properly wrapping other test rules in our own, we ensure
that these rules get the opportunity to reset themselves on retry,
and in particular that the Wiremock resets its expectations,
so that the second and third attempt start from a clean state.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>