HSEARCH-4172 Backport HSEARCH-4133 Remove the concept of "processing" in IndexIndexingPlan
The process() method was a no-op in both the Lucene backend and
Elasticsearch backend.

Its only use was for testing purposes, in the stub backend, to detect
that works have actually been created by the POJO mapper.
But this can be advantageously replaced with events sent when the
stub works are created.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>