HSEARCH-4133 Delay registering the HibernateSearchEventListener until the mapper starts
We're going to start the automatic indexing strategy then, so it's
easier to do everthing in the same place.

Technically, this also leads to a small change in behavior in
integrations that rely on an environment synchronizer to delay the
Hibernate Search boot.

Before this patch, we used to register the listener immediately, even
if the Hibernate Search boot was delayed. Any event triggered before
Hibernate Search had actually booted would be caught, but the listener
would block until Hibernate Search had booted.

The behavior is now different: we only register the listener when
Hibernate Search has booted. Any event triggered before Hibernate Search
has actually booted will be ignored.

I believe this is an acceptable behavior, considering that:

1. Environment synchronizers are only used in WildFly, which is not
   officially supported yet since it still ships ORM 5.3.
2. I doubt that an application that triggers entity changes before
   Hibernate Search has booted would work correctly anyway: it's likely
   that it would just deadlock because Hibernate Search blocks waiting for
   its boot, while WildFly waits for the entity changes to be processed
   before it proceeds with booting Hibernate Search.
3. At least the new behavior is safe in the sense that it won't lead
   to a deadlock.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>