HSEARCH-4369 Schedule outbox event processor tasks later when cluster link instructions require a delay
This should little effect except for the fact we will no longer poll
every 100ms when waiting until the pulse interval (around 2s) is elapsed;
that was completely pointless as the ScheduledThreadExecutor allows us
to just request for the worker to resume its work 2s later.

As an added benefit, IMO, the code is clearer because we spell out what
the delay will be when we schedule the worker for execution.