HSEARCH-4374 Remove LocalHeapQueueCooordinationStrategy
It's only used in tests, so this does not affect users.

This strategy was useful as a proof-of-concept when we didn't have the
outbox-polling strategy, but right now it's become a drain of resources,
both on CI (more tests executed)
and simply when maintaining (have to keep implementing new methods from
the CoordinationStrategy interface).