HSEARCH-4467 Adapt pom.xml for the checkstyle CI
The checkstyle CI needs to be able to run a newer version of checkstyle
against our project.

They used to override puppycrawl.checkstyle.version, and we had a hack
to make this property the default for
version.com.puppycrawl.tools.checkstyle, but that doesn't work well with
dependabot, so this will have to change.

Additionally, since their CI job was (unknowingly) still using an old
version of checkstyle, it didn't detect a problem that made the build
fail because of deprecations; I detected that now, and fixed it by
updating their CI job to not fail on compiler warnings.